// frontend/src/app/shared/openapi-api-client/model/accountRequest.ts
/**
 * scheduleme.server | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UpdateScheduleRequest } from './updateScheduleRequest';
import { UpdateUserFunctionRequest } from './updateUserFunctionRequest';

export interface AccountRequest {
  title?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  dob: string | null;
  schedules?: Array<UpdateScheduleRequest> | null;
  userFunctions?: Array<UpdateUserFunctionRequest> | null;
  role?: string | null;
  email?: string | null;
  password?: string | null;
  confirmPassword?: string | null;
  phoneNumber?: string | null;
}