// frontend/src/app/shared/openapi-api-client/model/createRequest.ts
/**
 * scheduleme.server | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateRequest {
  title: string | null;
  firstName: string | null;
  lastName: string | null;
  role: string | null;
  email: string | null;
  dob: string | null;
  phoneNumber?: string | null;
  password: string | null;
  confirmPassword: string | null;
}